+++ History.md
@@ -12,6 +12,7 @@
   * Your bugfix goes here (#Github Number)
   * Windows update extconf.rb for use with ssp and varied Ruby/MSYS2 combinations (#2069)
   * Preserve `BUNDLE_GEMFILE` env var when using `prune_bundler` (#1893)
+  * Send 408 request timeout even when queue requests is disabled (#2119)
 
 * Refactor
   * Remove unused loader argument from Plugin initializer (#2095)